Refactor big function into a helper
